Criando a camada Dao
