+++ egg/egg/Models/ItemDao.swift
@@ -1,0 +1,41 @@
+//
+//  ItemDao.swift
+//  egg
+//
+//  Created by Marco Aurelio on 30/10/19.
+//  Copyright © 2019 coda. All rights reserved.
+//
+
+import Foundation
+
+class ItemDao {
+    func save(_ itens: [Item]) {
+        do {
+            let dados = try NSKeyedArchiver.archivedData(withRootObject: itens, requiringSecureCoding: false)
+            guard let caminho = recuperaDiretorio() else { return }
+            try dados.write(to: caminho)
+        } catch {
+            print(error.localizedDescription)
+        }
+    }
+    
+    func recupera() -> [Item] {
+        do {
+            guard let diretorio = recuperaDiretorio() else { return [] }
+            let dados = try Data(contentsOf: diretorio)
+            let itensSalvos = try NSKeyedUnarchiver.unarchiveTopLevelObjectWithData(dados) as! [Item]
+            
+            return itensSalvos
+        } catch {
+            print(error.localizedDescription)
+            return []
+        }
+    }
+    
+    func recuperaDiretorio() -> URL? {
+        guard let diretorio = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else { return nil }
+        let caminho = diretorio.appendingPathComponent("itens")
+        
+        return caminho
+    }
+}
